First commit of astropy
from __doc__: This is the AstroPy module for writing astro libraries in Python. The way this works, is you register your function with the correct parameters, 
process some data in it and then exit the function to continue running the astro script. On the asx script side, you need a simple @mixin lib_function 
in order to execute the function with the current data in the scope.